Fixed compiling tests & fixed warnings (except THROWS macro)

+++ tests/src/QZXingTests/DecodeValidator.cpp
@@ -1,8 +1,10 @@
 #include "DecodeValidator.h"
 
+#include <QDebug>
 #include <QFile>
 #include <QDirIterator>
 #include <QFileInfo>
+#include <QTextStream>
 
 #define LOG_OUTPUT_DIVIDER  "##############################################"
 #define LOG_SECTOR_TITLE(a)    '\n' << LOG_OUTPUT_DIVIDER\
@@ -123,13 +125,13 @@ void DecodeValidator::printResults()
         if(failedResultLogs.size())
             qDebug() << "  failed image files:";
 
-        for(size_t i=0; i<failedResultLogs.size(); i++)
+        for(int i=0; i<failedResultLogs.size(); i++)
             qDebug() << '\t' << failedResultLogs[i];
 
         if(inconsistentResultLogs.size())
             qDebug() << "  inconsistent image files:";
 
-        for(size_t i=0; i<inconsistentResultLogs.size(); i++)
+        for(int i=0; i<inconsistentResultLogs.size(); i++)
             qDebug() << '\t' << inconsistentResultLogs[i];
     }
